完成目录页：悬浮头 一级目录、二级目录 阅读页，完成基本功能 自定义Toast

+++ app/src/main/java/com/novel/qingwen/bean/ChapterContent.kt
@@ -7,12 +7,12 @@ data class ChapterContent(
 )
 
 data class ChapterData(
-    val cid: Int,
+    val cid: Long,
     val cname: String,
     val content: String,
     val hasContent: Int,
-    val id: Int,
+    val id: Long,
     val name: String,
-    val nid: Int,
-    val pid: Int
+    val nid: Long,
+    val pid: Long
 )
